Kodland_BOT/requirements.txt
discord.py
sqlite3
pytest